Refactor symbol package acquisition boundaries
