+++ src/DotnetInspector.Packages/SymbolPackageDownloader.cs
@@ -1,6 +1,4 @@
-using System.Net;
 using DotnetInspector.Core;
-using InertText;
 
 namespace DotnetInspector.Packages;
 
@@ -116,7 +114,7 @@ public abstract record PortablePdbAcquisitionResult
 /// Portable PDBs use {GUID}FFFFFFFF, Windows PDBs use {GUID}{age:x}.
 /// </para>
 /// </remarks>
-public class SymbolPackageDownloader
+public partial class SymbolPackageDownloader
 {
     private const string SymbolMissCacheCategory = "symbol-misses";
     private static readonly TimeSpan SymbolMissCacheTtl = TimeSpan.FromDays(1);
@@ -344,539 +342,6 @@ public class SymbolPackageDownloader
             acquired.Pdb.SymbolServer);
     }
 
-    private bool IsNuGetOrgEligibleForPackage(
-        NuGetSourceOptions? sourceOptions,
-        string packageName)
-    {
-        if (_sourceAuthorization is not null)
-        {
-            PackageSourceAuthorization authorization =
-                _sourceAuthorization.AuthorizeSourcesFor(
-                    packageName.ToLowerInvariant());
-            return authorization.Sources.Any(
-                source => source.IsNuGetOrg);
-        }
-
-        try
-        {
-            return NuGetSourceResolver.ResolveSourcesForPackage(
-                    sourceOptions,
-                    packageName)
-                .Any(source => source.IsNuGetOrg);
-        }
-        catch (PackageSourceMappingException ex)
-            when (ex.Failure is
-                PackageSourceMappingFailure.NoPattern
-                or PackageSourceMappingFailure.InactiveSource)
-        {
-            return false;
-        }
-    }
-
-    private sealed record PdbProbeResult(
-        AcquiredPortablePdb? Pdb,
-        bool WindowsPdbDetected);
-
-    private PdbProbeResult Acquired(
-        string cacheKey,
-        string symbolServer,
-        bool fromCache,
-        bool windowsPdbDetected = false)
-        => new(
-            new AcquiredPortablePdb(
-                _pdbStore,
-                cacheKey,
-                symbolServer,
-                fromCache),
-            windowsPdbDetected);
-
-    private async Task<PdbProbeResult> TryLocateFromMsdlAsync(
-        string pdbFileName,
-        string symbolKey,
-        string storeIdentity,
-        Guid pdbGuid,
-        uint? portablePdbStamp,
-        bool isPortable,
-        Action<string>? log,
-        bool cacheOnly,
-        CancellationToken cancellationToken)
-    {
-        using var trafficScope = NetworkTelemetry.Scope(NetworkTrafficKind.SymbolDownload);
-        const string ServerHost = "msdl.microsoft.com";
-        bool windowsPdbDetected = false;
-
-        var cacheKey =
-            GetSymbolServerCacheKey(
-                ServerHost,
-                pdbFileName,
-                storeIdentity);
-        var cached = await ClassifyStoredPdbAsync(
-            cacheKey,
-            pdbGuid,
-            portablePdbStamp,
-            isPortable,
-            log,
-            cancellationToken).ConfigureAwait(false);
-        if (cached.Portable)
-        {
-            log?.Invoke("Using cached PDB from MSDL");
-            return Acquired(
-                cacheKey,
-                ServerHost,
-                fromCache: true);
-        }
-        if (cached.Windows)
-            windowsPdbDetected = true;
-
-        if (cacheOnly)
-            return new PdbProbeResult(null, windowsPdbDetected);
-
-        var url = $"https://msdl.microsoft.com/download/symbols/{pdbFileName}/{symbolKey}/{pdbFileName}";
-        if (IsCachedMiss(url, log, "MSDL symbol server"))
-            return new PdbProbeResult(null, windowsPdbDetected);
-
-        log?.Invoke("Trying MSDL symbol server");
-
-        bool storeOperation = false;
-        try
-        {
-            var httpResult = await HttpRetryHelper.GetWithRetryResultAsync(
-                _client, url, log: log,
-                cancellationToken: cancellationToken,
-                trafficKind: NetworkTrafficKind.SymbolDownload).ConfigureAwait(false);
-            using var response = httpResult.Response;
-            if (response == null || !response.IsSuccessStatusCode)
-            {
-                CacheMissIfDefinitive(url, httpResult);
-                log?.Invoke("MSDL: symbol not found");
-                return new PdbProbeResult(null, windowsPdbDetected);
-            }
-
-            using (var content = await response.Content.ReadAsStreamAsync(cancellationToken).ConfigureAwait(false))
-            {
-                storeOperation = true;
-                await _pdbStore.PutAsync(cacheKey, content, cancellationToken).ConfigureAwait(false);
-            }
-
-            var headerCheck = await ClassifyStoredPdbAsync(
-                cacheKey,
-                pdbGuid,
-                portablePdbStamp,
-                isPortable,
-                log,
-                cancellationToken).ConfigureAwait(false);
-            storeOperation = false;
-            if (headerCheck.Portable)
-            {
-                log?.Invoke("Successfully downloaded PDB from MSDL");
-                return Acquired(
-                    cacheKey,
-                    ServerHost,
-                    fromCache: false,
-                    windowsPdbDetected);
-            }
-            if (headerCheck.Windows)
-            {
-                windowsPdbDetected = true;
-                log?.Invoke("MSDL returned a Windows PDB (not supported)");
-            }
-        }
-        catch (OperationCanceledException) when (cancellationToken.IsCancellationRequested)
-        {
-            throw;
-        }
-        catch (Exception ex) when (!storeOperation)
-        {
-            log?.Invoke($"MSDL error: {ex.Message}");
-        }
-
-        return new PdbProbeResult(null, windowsPdbDetected);
-    }
-
-    private async Task<PdbProbeResult> TryLocateFromSymbolPackageAsync(
-        string packageName,
-        string packageVersion,
-        string assemblyName,
-        string symbolKey,
-        string storeIdentity,
-        Guid pdbGuid,
-        uint? portablePdbStamp,
-        bool isPortable,
-        Action<string>? log,
-        bool cacheOnly,
-        CancellationToken cancellationToken)
-    {
-        using var trafficScope = NetworkTelemetry.Scope(NetworkTrafficKind.SymbolDownload);
-        var normalizedName = packageName.ToLowerInvariant();
-        var normalizedVersion = packageVersion.ToLowerInvariant();
-        bool windowsPdbDetected = false;
-
-        // Check cache first
-        var cacheKey =
-            GetCachedPdbKey(
-                normalizedName,
-                normalizedVersion,
-                assemblyName,
-                storeIdentity);
-        var cached = await ClassifyStoredPdbAsync(
-            cacheKey,
-            pdbGuid,
-            portablePdbStamp,
-            isPortable,
-            log,
-            cancellationToken).ConfigureAwait(false);
-        if (cached.Portable)
-        {
-            log?.Invoke($"Using cached PDB: {assemblyName}.pdb");
-            return Acquired(
-                cacheKey,
-                "nuget.org",
-                fromCache: true);
-        }
-        if (cached.Windows)
-            windowsPdbDetected = true;
-
-        if (cacheOnly)
-            return new PdbProbeResult(null, windowsPdbDetected);
-
-        // Try NuGet global CDN first
-        var snupkgUrls = new[]
-        {
-            $"https://globalcdn.nuget.org/symbol-packages/{normalizedName}.{normalizedVersion}.snupkg",
-            $"https://api.nuget.org/v3-flatcontainer/{normalizedName}/{normalizedVersion}/{normalizedName}.{normalizedVersion}.snupkg"
-        };
-
-        log?.Invoke($"Trying symbol package: {normalizedName}.{normalizedVersion}.snupkg");
-
-        foreach (var snupkgUrl in snupkgUrls)
-        {
-            if (IsCachedMiss(snupkgUrl, log, "symbol package"))
-                continue;
-
-            HttpRetryHelper.HttpRetryResult httpResult;
-            try
-            {
-                httpResult = await HttpRetryHelper.GetWithRetryResultAsync(
-                    _client, snupkgUrl, log: log,
-                    cancellationToken: cancellationToken,
-                    trafficKind: NetworkTrafficKind.SymbolDownload).ConfigureAwait(false);
-            }
-            catch (OperationCanceledException) when (cancellationToken.IsCancellationRequested)
-            {
-                throw;
-            }
-            catch (Exception ex)
-            {
-                log?.Invoke(
-                    "Error downloading symbol package: "
-                    + UrlRedaction.DescribeRequestFailure(snupkgUrl, ex));
-                continue;
-            }
-
-            using var response = httpResult.Response;
-            if (response is not { IsSuccessStatusCode: true })
-            {
-                CacheMissIfDefinitive(snupkgUrl, httpResult);
-                continue;
-            }
-
-            SnupkgPdbResult extracted;
-            try
-            {
-                log?.Invoke(
-                    $"Found symbol package at: {UrlRedaction.ForDiagnostics(snupkgUrl)}");
-                extracted = await ExtractPdbFromSymbolPackage(
-                    response,
-                    assemblyName,
-                    pdbGuid,
-                    portablePdbStamp,
-                    log,
-                    cancellationToken).ConfigureAwait(false);
-            }
-            catch (OperationCanceledException) when (cancellationToken.IsCancellationRequested)
-            {
-                throw;
-            }
-            catch (Exception ex)
-            {
-                log?.Invoke(
-                    "Error reading symbol package: "
-                    + UrlRedaction.DescribeRequestFailure(snupkgUrl, ex));
-                continue;
-            }
-
-            if (extracted.WindowsPdbDetected)
-                windowsPdbDetected = true;
-
-            if (extracted.PdbBytes == null)
-            {
-                log?.Invoke(
-                    "No matching Portable PDB identity found in symbol package");
-                return new PdbProbeResult(
-                    null,
-                    windowsPdbDetected);
-            }
-
-            using (var pdbStream =
-                   new MemoryStream(
-                       extracted.PdbBytes,
-                       writable: false))
-            {
-                await _pdbStore.PutAsync(
-                    cacheKey,
-                    pdbStream,
-                    cancellationToken).ConfigureAwait(false);
-            }
-
-            var stored =
-                await ClassifyStoredPdbAsync(
-                    cacheKey,
-                    pdbGuid,
-                    portablePdbStamp,
-                    isPortable,
-                    log,
-                    cancellationToken).ConfigureAwait(false);
-            if (!stored.Portable)
-            {
-                if (stored.Windows)
-                    windowsPdbDetected = true;
-                log?.Invoke(
-                    "The matching Portable PDB could not be read back from the configured store.");
-                return new PdbProbeResult(
-                    null,
-                    windowsPdbDetected);
-            }
-
-            log?.Invoke(
-                "Successfully located PDB from symbol package");
-            return Acquired(
-                cacheKey,
-                "nuget.org",
-                fromCache: false,
-                windowsPdbDetected);
-        }
-
-        log?.Invoke("Symbol package not found on NuGet");
-        return new PdbProbeResult(null, windowsPdbDetected);
-    }
-
-    private static async Task<SnupkgPdbResult> ExtractPdbFromSymbolPackage(
-        HttpResponseMessage response,
-        string assemblyName,
-        Guid pdbGuid,
-        uint? portablePdbStamp,
-        Action<string>? log,
-        CancellationToken cancellationToken)
-    {
-        SnupkgPdbResult extracted;
-        using (var content = await response.Content.ReadAsStreamAsync(cancellationToken).ConfigureAwait(false))
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-            extracted = SnupkgPdbReader.ExtractPortablePdb(
-                content,
-                assemblyName,
-                pdbGuid,
-                log,
-                portablePdbStamp);
-            cancellationToken.ThrowIfCancellationRequested();
-        }
-
-        return extracted;
-    }
-
-    private async Task<PdbProbeResult> TryLocateFromSymbolServerAsync(
-        string pdbFileName,
-        string symbolKey,
-        string storeIdentity,
-        Guid pdbGuid,
-        uint? portablePdbStamp,
-        bool isPortable,
-        Action<string>? log,
-        bool cacheOnly,
-        CancellationToken cancellationToken)
-    {
-        using var trafficScope = NetworkTelemetry.Scope(NetworkTrafficKind.SymbolDownload);
-        bool windowsPdbDetected = false;
-
-        var symbolServers = new[]
-        {
-            "https://symbols.nuget.org/download/symbols",
-            "https://msdl.microsoft.com/download/symbols"
-        };
-
-        foreach (var server in symbolServers)
-        {
-            var serverHost = new Uri(server).Host;
-            var cacheKey =
-                GetSymbolServerCacheKey(
-                    serverHost,
-                    pdbFileName,
-                    storeIdentity);
-            var cached =
-                await ClassifyStoredPdbAsync(
-                    cacheKey,
-                    pdbGuid,
-                    portablePdbStamp,
-                    isPortable,
-                    log,
-                    cancellationToken).ConfigureAwait(false);
-            if (cached.Portable)
-            {
-                log?.Invoke(
-                    $"Using cached PDB from {serverHost}");
-                return Acquired(
-                    cacheKey,
-                    serverHost,
-                    fromCache: true,
-                    windowsPdbDetected);
-            }
-            if (cached.Windows)
-                windowsPdbDetected = true;
-
-            if (cacheOnly)
-                continue;
-
-            var url = $"{server}/{pdbFileName}/{symbolKey}/{pdbFileName}";
-            if (IsCachedMiss(url, log, "symbol server"))
-                continue;
-
-            log?.Invoke($"Trying symbol server: {server}");
-
-            bool storeOperation = false;
-            try
-            {
-                var httpResult = await HttpRetryHelper.GetWithRetryResultAsync(
-                    _client, url, log: log,
-                    cancellationToken: cancellationToken,
-                    trafficKind: NetworkTrafficKind.SymbolDownload).ConfigureAwait(false);
-                using var response = httpResult.Response;
-                if (response == null || !response.IsSuccessStatusCode)
-                {
-                    CacheMissIfDefinitive(url, httpResult);
-                    continue;
-                }
-
-                using (var content = await response.Content.ReadAsStreamAsync(cancellationToken).ConfigureAwait(false))
-                {
-                    storeOperation = true;
-                    await _pdbStore.PutAsync(cacheKey, content, cancellationToken).ConfigureAwait(false);
-                }
-
-                var headerCheck = await ClassifyStoredPdbAsync(
-                    cacheKey,
-                    pdbGuid,
-                    portablePdbStamp,
-                    isPortable,
-                    log,
-                    cancellationToken).ConfigureAwait(false);
-                storeOperation = false;
-                if (headerCheck.Portable)
-                {
-                    log?.Invoke("Successfully downloaded PDB from symbol server");
-                    return Acquired(
-                        cacheKey,
-                        serverHost,
-                        fromCache: false,
-                        windowsPdbDetected);
-                }
-                if (headerCheck.Windows)
-                {
-                    windowsPdbDetected = true;
-                    log?.Invoke("Symbol server returned a Windows PDB (not supported)");
-                }
-            }
-            catch (OperationCanceledException) when (cancellationToken.IsCancellationRequested)
-            {
-                throw;
-            }
-            catch (Exception ex) when (!storeOperation)
-            {
-                log?.Invoke($"Symbol server error: {ex.Message}");
-            }
-        }
-
-        return new PdbProbeResult(null, windowsPdbDetected);
-    }
-
-    private async Task<(bool Portable, bool Windows)> ClassifyStoredPdbAsync(
-        string cacheKey,
-        Guid expectedGuid,
-        uint? expectedStamp,
-        bool expectedPortable,
-        Action<string>? log,
-        CancellationToken cancellationToken)
-    {
-        Stream? stream;
-        try
-        {
-            stream = await _pdbStore.TryOpenAsync(cacheKey, cancellationToken).ConfigureAwait(false);
-        }
-        catch (ArgumentException)
-        {
-            // An untrusted assembly/PDB name can produce a key the store rejects;
-            // that simply means nothing is (or can be) cached under it.
-            return (false, false);
-        }
-
-        if (stream == null)
-            return (false, false);
-
-        await using (stream.ConfigureAwait(false))
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-            var header = SnupkgPdbReader.ClassifyHeader(stream);
-            if (!header.Portable || !expectedPortable)
-                return (false, header.Windows);
-
-            stream.Position = 0;
-            bool matches =
-                SnupkgPdbReader.PortablePdbMatchesIdentity(
-                    stream,
-                    expectedGuid,
-                    expectedStamp,
-                    log);
-            return (matches, false);
-        }
-    }
-
-    private bool IsCachedMiss(
-        string key,
-        Action<string>? log,
-        string source)
-    {
-        if (!_usePersistentMissCache)
-            return false;
-
-        if (CoreCache.TryGet(SymbolMissCacheCategory, key, SymbolForbiddenCacheTtl, extension: "forbidden") != null)
-        {
-            log?.Invoke($"Using cached symbol miss: {source}");
-            return true;
-        }
-
-        if (CoreCache.TryGet(SymbolMissCacheCategory, key, SymbolMissCacheTtl, extension: "miss") == null)
-            return false;
-
-        log?.Invoke($"Using cached symbol miss: {source}");
-        return true;
-    }
-
-    private void CacheMissIfDefinitive(
-        string key,
-        HttpRetryHelper.HttpRetryResult result)
-    {
-        if (!_usePersistentMissCache)
-            return;
-
-        if (result.StatusCode is not { } statusCode)
-            return;
-
-        if (HttpRetryHelper.IsRetryableStatus(statusCode))
-            return;
-
-        var extension = statusCode == HttpStatusCode.Forbidden ? "forbidden" : "miss";
-        CoreCache.Set(SymbolMissCacheCategory, key, ((int)statusCode).ToString(), extension);
-    }
-
     /// <summary>
     /// Checks if a package name indicates it's a Microsoft package.
     /// </summary>
@@ -890,19 +355,6 @@ public class SymbolPackageDownloader
                packageName.StartsWith("Azure.", StringComparison.OrdinalIgnoreCase) ||
                packageName.Equals("WindowsAzure.Storage", StringComparison.OrdinalIgnoreCase);
     }
-
-    private static string GetCachedPdbKey(
-        string packageName,
-        string packageVersion,
-        string assemblyName,
-        string symbolKey)
-        => $"{packageName}/{packageVersion}/{symbolKey}/{assemblyName}.pdb";
-
-    private static string GetSymbolServerCacheKey(
-        string serverHost,
-        string pdbName,
-        string symbolKey)
-        => $"servers/{serverHost}/{pdbName}/{symbolKey}/{pdbName}";
 
     private static string? GetSnupkgAssemblyName(
         string pdbFileName,
